Create Arduino_Sim_V2_working
Version 2 which is in progress and when a switch is turned on it runs it once and only runs the other command once the switch is turned off. Instead of repeating the command over and over again. This is for a simulator ex: landing gear down= off, landing gear up = on